Move data.js into /helpers
