+++ app/(pages)/trending/top10/page.jsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import { Table, TableCell, TableContainer, TableHead, TableRow, Paper } from '@mui/material';
 import ScoreTableBody from '@/components/ScoreTableBody';
-import { top10Nasdaq } from '@/app/data';
+import { top10Nasdaq } from '@/helpers/data';
 import { Box, Container } from '@mui/joy';
 import TitlePanel from '@/components/TitlePanel';
