+++ app/(pages)/trending/top10gov/page.jsx
@@ -4,7 +4,7 @@ import TitlePanel from "@/components/TitlePanel";
 import { fetchNasdaq100Category } from "@/helpers/requests";
 import { Box, Container } from "@mui/joy";
 import { useState, useEffect } from "react";
-import { govOpp10, govRisk10, trendingDesc } from "@/app/data";
+import { govOpp10, govRisk10, trendingDesc } from "@/helpers/data";
 import OppRiskDisplay from "@/components/OppRiskDisplay";
 
 export default function () {
